+++ changes.md
@@ -6,3 +6,10 @@
 - :warning: changed property `actions` of type `MyBusinessUnitUpdate` from type `BusinessUnitUpdateAction[]` to `MyBusinessUnitUpdateAction[]`
 </details>
 
+
+<details>
+<summary>Added Property(s)</summary>
+
+- added property `published` to type `ShoppingListLineItem`
+</details>
+